+++ bittensor/core/addons/subtensor_api/utils.py
@@ -1,7 +1,7 @@
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from bittensor.core.subtensor_api import SubtensorApi
+    from bittensor.core.addons import SubtensorApi
 
 
 def add_legacy_methods(subtensor: "SubtensorApi"):
